test case refactor
+ testcase_helper.rb module
