+++ app/models/student_test_case.rb
@@ -2,5 +2,7 @@ class StudentTestCase < ActiveRecord::Base
 
     # Relationships
     belongs_to :coding_prompt_answer
+    
+    #refactored to testcase
 
 end
